Address Greptile report review
